refactor: move purchaseStorageExpansion to its own file
