+++ src/game-logic/reducers/index.test.js
@@ -9,9 +9,8 @@ import {
   MAX_ANIMAL_NAME_LENGTH,
   MAX_LATEST_PEER_MESSAGES,
   MAX_PENDING_PEER_MESSAGES,
-  STORAGE_EXPANSION_AMOUNT,
 } from '../../constants'
-import { generateCow, getCostOfNextStorageExpansion } from '../../utils'
+import { generateCow } from '../../utils'
 
 import * as fn from './'
 
@@ -35,32 +34,6 @@ describe('selectCow', () => {
   test('updates selectedCowId', () => {
     const { selectedCowId } = fn.selectCow({}, { id: 'abc' })
     expect(selectedCowId).toEqual('abc')
-  })
-})
-
-describe('purchaseStorageExpansion', () => {
-  describe('player does not have enough money', () => {
-    test('no-ops', () => {
-      const inputState = {
-        money: getCostOfNextStorageExpansion(100) - 1,
-        inventoryLimit: 100,
-      }
-
-      const state = fn.purchaseStorageExpansion(inputState)
-      expect(state).toEqual(inputState)
-    })
-  })
-
-  describe('player has enough money', () => {
-    test('increase storage is purchased', () => {
-      const { inventoryLimit, money } = fn.purchaseStorageExpansion({
-        inventoryLimit: 100,
-        money: getCostOfNextStorageExpansion(100),
-      })
-
-      expect(inventoryLimit).toEqual(100 + STORAGE_EXPANSION_AMOUNT)
-      expect(money).toEqual(0)
-    })
   })
 })
